Fix theme file collection for duplicate names (#86)
* fix theme file collection for duplicate names

WP_Theme::get_files() returns an associative array with the relative
file path as key and the absolute path as value. For child themes
with files matching the parent's relative paths, these got overwritten
by array_merge(). We only process the absolute path, so now we strip
the values and use numeric indices which are not squashed.

* remove default value from line content