drop recursive check for get_option() calls in theme check
Antivirus looks for get_option() calls in theme files and executes a
recursive iteration on the string representation of the value retrieved.

While the real use of this (old) check is unclear, there are several
scenarios where the logic breaks our plugin. One is the broken iteration
when the value is of array type (notice in 7.4, warning in 8.0). Another
more severe problem might arise from options containing strings with
"get_option( ... )" calls itself, causing infinite recursion that ends
up in a stack overflow.

Because of the dubious benefits, we drop this check completely for now.