mark file as OK only if all issues have been dismissed (#135) (#136)
When multiple issues are present in a single file, clicking on "dismiss"
should only hide the corresponding line and the file should only be
marked as "OK" and turned green, if there are no more issues present.

Refactor the response handler and only remove the parent element to
resolve this issue.