do not encode URLs in SafeBrowsing request (#156)
The urlencode() call is a leftover from the old SafeBrowsing v3 check.
The v4 API accepts a list of URLs as JSON in the request body where no
special encoding is required.

Apparently it does not break the check as the API seems to accept both,
but it is not necessary at this point. Remove it.