fix: preserve event listener on multiple dismiss buttons (#161) (#162)
Appending to innerHTML makes the browser re-render the parent node which
removes the previously added event listener.
Because of this only the last "dismiss" button within a file row
actually works.

Rewrite the logic to create a node programmatically and append it, so
the element does not get reset in the loop.

Fixes: 819272cef9476cc1eeb66bfcaa744f5e07a4adf3